Rename brain module to core
